User authentication

+++ todo-app/models/todo.js
@@ -1,5 +1,8 @@
+const { Op } = require("sequelize");
+
+/* eslint-disable no-undef */
 /* eslint-disable no-unused-vars */
-"use strict";
+("use strict");
 const { Model } = require("sequelize");
 module.exports = (sequelize, DataTypes) => {
 	class Todo extends Model {
@@ -12,17 +15,48 @@ module.exports = (sequelize, DataTypes) => {
 			// define association here
 			Todo.belongsTo(models.User, { foreignKey: "userId" });
 		}
-		static addTodo({ title, dueDate }) {
-			return this.create({ title, dueDate, completed: false });
+		static addTodo({ title, dueDate, userId }) {
+			return this.create({ title, dueDate, completed: false, userId });
 		}
-		setCompletionStatus() {
+		static setCompletionStatus() {
 			return this.update({ completed: !this.completed });
 		}
-		static removeTodo(id) {
-			return this.destroy({ where: { id } });
+		static removeTodo(id, userId) {
+			return this.destroy({ where: { id: id, userId: userId } });
 		}
+
 		static getTodos() {
 			return this.findAll();
+		}
+		static getOverdue(userId) {
+			return this.findAll({
+				where: {
+					dueDate: { [Op.lt]: new Date() },
+					completed: false,
+					userId,
+				},
+			});
+		}
+		static getDueToday(userId) {
+			return this.findAll({
+				where: {
+					dueDate: { [Op.eq]: new Date() },
+					completed: false,
+					userId,
+				},
+			});
+		}
+		static getDueLater(userId) {
+			return this.findAll({
+				where: {
+					dueDate: { [Op.gt]: new Date() },
+					completed: false,
+					userId,
+				},
+			});
+		}
+		static getCompleted(userId) {
+			return this.findAll({ where: { completed: true, userId } });
 		}
 	}
 	Todo.init(
